refactor(ProgressStepBar): extract and export shared types via dedicated module
