+++ src/components/ProgressStepBar/index.ts
@@ -1,1 +1,2 @@
 export * from './ProgressStepBar';
+export * from './types';
